Not require trailing empty space in label test
The test for ensuring a LABEL can have a multi line value
requires the '\' to be present at end of line to indicate
value continues to next line. A trailing white space was
used to force the python compiler into assuming the '\'
was part of the string. A better approach is to escape
the character.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>