Combine environment variable substitution with quote removal
This fixes the tests added in the previous commit.

The EnvSubst class is renamed to WordSplitter, and provides
two methods:
* dequote() returns the string with quotes/escaped consumed
* split() breaks the string into words at whitespace

This new class can be told whether or not to apply
environment variable substitutions.

The way we process ENV/LABEL instructions is now to first
break the instruction value into words, without quoting or
applying environment variable substitutions. We can then
inspect the first word to find out which key/value form is
in use ("key=value" vs "key the value").

Based on this we now apply environment variable substitution
to each key and value separately, consuming quotes and
escape characters.

A similar process is used when modifying an ENV/LABEL
instruction, with the exception that key=value forms are
left alone with quotes and escapes intact unless modified.
This is the reason for changing some existing test cases.

Signed-off-by: Tim Waugh <twaugh@redhat.com>