Handle escape directive
This change scans the beginning of the Dockerfile for an escape directive as described in the docker Documentation. In case this is set the regex for the line continuation and the newly introduced state for the line continuation character is set. In addition the _rstrip_backslash was renamed to _rstrip_eol and takes one additional argument for the line continuation character.

Signed-off-by: Thomas Scheruebl <thomas.scheruebl@intel.com>